microservices/resume_parser/requirements.txt
fastapi
uvicorn
python-dotenv
boto3
transformers
torch
sentencepiece
python-docx
pymupdf
regex
pydantic
python-multipart
